Refactor diff format diff utils
diff_format was getting hard to navigate due to its size and many
functions. This extracts some of the funtionality into a separate file
diff_utils. The guiding principle is supposed to be a separation of
concerns, where functions that construct/verify diff entries are in the
diff_format file, and everything else goes into utils.
